src/server/requirements.txt
web.py==0.39
ast2json==0.2.1
pymongo==3.7.2
PyJWT==1.6.4